refactor: 최근 회원 코멘트 조회 UI 및 네트워크 로직 개선
- 회원 코멘트 응답 모델 확장 및 comment 필드명 반영
- 최근 코멘트 셀 UI 분리 및 닉네임·성별·나이·하트·거리 표시
- 커서 기반 댓글 조회 API 연동 및 무한 스크롤·당겨서 새로고침
- UserDetailView id 타입을 Int64로 변경하고 호출부 일괄 수정
- 상대 시간 표현용 DateHelper 유틸 추가
- 회원 차단·하트·비밀사진·검색 리스트의 네비게이션 링크 타입 정리
- 회원 활성화 화면에 공통 알림(Alert) 처리 추가
- 인증 인터셉터의 401/403/500 처리 분기 보완 및 안정성 개선

+++ app/gugu/BlockListView.swift
@@ -9,7 +9,7 @@ struct BlockListView: View {
                 ScrollView {
                     LazyVStack(alignment: .leading) {
                         ForEach(0..<100) { i in
-                            NavigationLink(destination: UserDetailView(id: i)) {
+                            NavigationLink(destination: UserDetailView(id: Int64(i))) {
                                 HStack(alignment: .center) {
                                     if i.isMultiple(of: 2) {
                                         AsyncImage(url: URL(string: "https://picsum.photos/100")) { phase in
